Initial commit of Active Directory functionality.
This adds utilities and constants used to format LDAP attributes,
parse the values of AD attributes read via LDAP, define kerberos
encryption types in AD, and translate them to normal kerberos
contexts. These utilities can also be used to construct and ingest
kerberos keys.

It also adds structural classes that define:
- raw kerberos keys, tied only to a password and encryption type
- complete kerberos keys, which define things such as principal and kvno
  that are needed to initiate or accept GSS contexts
- Active Directory domains
- Active Directory computers
- Active Directory sessions, defined by a domain and an LDAP connection to it

The domains provide useful functionality for automatically discovering
quick-to-respond LDAP and Kerberos servers on the network using the
properties of AD in DNS. They also provide functionality for establishing
sessions with domains given credentials, authentication mechanism, and
other connection attributes.
This largely builds on the ldap3 library, with some augmentation based
on quirks with AD such as differing authentication restrictions for users
vs. computers.

The sessions provide helper functions for creating computers, checking if
an object exists in a domain with a certain attribute, and checking if a
DN exists.

The key piece of functionality defined at this time though are the helper
functions for joining a domain.
Simple functions, which can be called with minimal information, allow
a script to join a computer to the AD domain. This process will create
a computer to represent the computer that the script is running on
within the domain, and will automatically generate kerberos keys for
the machine and write them to the default kerberos keytab file.

This allows a script to very easily, and without intimate AD knowledge,
configure the computer such that any other process or script on the
computer will be able to both accept GSSAPI security contexts
(e.g. authenticate incoming ssh sessions or SMB mounts using kerberos)
and initiate GSSAPI security contexts (e.g. establish kerberos
authenticated LDAP sessions, HTTP connections, etc.).

By default, the computer and keys created in the domain are incredibly
secure, and follow all of the naming and service conventions that are
standard in AD using the local computer's hostname.
However, things like services, hostnames, encryption types, etc. can
all also be customized. The join function is also incredibly flexible
so that a power user can customize currently unsupported attributes to
a large extent.

Planned future functionality includes:
- support for more options in joining, such as uploading machine
  certificates to bootstrap mTLS
- support for bootstrapping trust with the domain by reading in
  the certs for the CA in the domain
- support for creating users and groups, finding them, and managing
  their memberships
- support for site-based AD domains instead of just base domain
- support for discovering trusted domains, and better multi-domain
  and subdomain support
- better support within the library for configuring system kerberos
  settings. Currently all configuration of things like the local
  krb5.conf must be done outside of this library