Added site-awareness support to AD domain definitions.
Made it so that when joining an AD domain or defining an AD domain,
a site can be specified in order to ensure that communication
only happens with controllers in that site.

This allows people using the module to confine communication to
controllers within a site, which may be necessary in order to
ensure tight synchronization (for example, controllers in a site
might be kept consistent in terms of records while they can be
up to a few hours out of sync with other sites' controllers)
and avoid inconsistent behavior during failovers across servers.
It can also be used to make finding servers to talk to more
efficient if network firewalls/vlan configurations keep the
host the module is running isolated and prevent it from talking
to servers outside of a specific site.

Additionally, type annotations were added in many more places,
and docstrings were enhanced.