Added custom DNS and source address support.
Fleshed out some docstrings as well.
Added support for specifying a list of dns nameservers for the domain.
This allows server discovery in dns even when the system is pointing
at different dns nameservers; for example, when joining a new workstation
to a domain, it might be pointed at default ISP dns servers and you don't
want to de-configure those yet, so you just set the domain's dns servers
for join time.

Added support for setting a source IP to use when interacting with the
domain. This will be used for dns lookups, ldap connections in sessions,
and checking reachability/rtt to ldap and kerberos servers.

The above 2 enhancements are very useful for containerized environments
and multi-tenant network setups, as a single machine may have multiple
identities on a network and it's important that scripts be able to
choose which one they're operating as. Underlying system networking on
most OS's is not sufficiently flexible/easy to configure for us to
always rely on it to figure out what we want to do based on what
we're talking to.