Added functionality for getting domain-level functional attributes.
There's now functions for getting the time in the domain and the
supported sasl mechanisms. There's also one for getting the domain
functional level.
Enums have been added for functional level.

There's also a function for checking if the domain time is close to
the local time, with "close" defaulting to 5 minutes (which is
the default for kerberos)