Added functionality for retrieving DNS servers and CA certificates.
Also includes some bug fixes for the last commit around getting time
and functional level.

There's now functions to find the certificate authorities within
an AD domain, and return all of their CA certificates in a list,
with options to get them in DER or PEM format.
There's also a function to find all computers in the domain with
DNS as a service registered on them, and attempt to map them
to IP addresses. It then returns a list mapping the hostnames of
the dns servers to their ip addresses, enabling the caller to
configure them for plain dns or dns over tls.