Added membership update functionality.
The following functions have been added:
- add users to groups
- add groups to groups
- remove users from groups
- remove groups from groups

The functions attempt to be idempotent. If a user or group is already
not a member of a group, then it will not be remove. If it's already
a member, it will not be added. So no errors will arise from that.
Additionally, by default, if an error occurs updating a group, we
attempt to rollback all the changes made so far.

If rollback fails, an error is raised with information on the potentially
still modified groups. If it succeeds, an error is raised indicating that
rollback succeeded, and which group failed to update and triggered it.

There's also an option to not rollback, and just continue skipping the
groups that cause issued.

A list of successfully updated groups is returned.