Corrected an issue around duplicate SRV records.
Active Directory has an issue where older versions registered their
domain controllers' SRV records using the domain controller name
as is, which can include capital letters, but newer versions check
for records after casting names to lowercase.

The result is registration of duplicate DNS records for the same
domain controllers during upgrades for some domains, which must be
manually corrected.

DNS names are case insensitive. To make the library play nicer with
domains that do not remedy these duplicate, our discovery utils now
filter out duplicates before sorting by RTT. This saves time that
would be spent checking duplicates, and also makes it so that if
we have a server_limit set, we actually get server_limit unique
hostname:port uris in our response.

The function for querying dns does not itself do the de-duping, in
case users want to use that to check for duplicates (though it's a
private function).

Shoutout to @reach4bawer for running into this and filing the first
ever issue against this package!