Added security descriptor retrieval functionality, LDAP controls, and paging.
Added the ability to read the security descriptor for either a user, a group,
or any generic AD object. This builds on the parsing functionality that was
previously added.

Added support for specifying LDAP controls in ALL search-type operations that
the library supports for finding things that are not domain-level attributes.

Added support for auto-paging. The default threshold is 100 results, and it
can be set.
The idea is that searches that have the possibility to exceed the page size will
use a paginated search. Searches that have lower search limits will use a normal
search. This makes our big searches friendlier in a multi-threaded application.