Fixed up some bugs and imports.
Fixed an issue where paginated responses were in a different format
than non-paginated, so unpacking needed to happen differently.

Searching by SID should use a size_limit of 1 for searches, and
that was not being done.
Size limit was added as a parameter to searching by attribute,
since it could be useful.

Some enum values were renamed, and the enum was made available
at the top level.