Added functions for setting the security descriptor on a user, group, or object.
Added 3 new functions to set the security descriptor for users,
groups, or generic objects.
These support changing the seucirty descriptor in a variety of
ways when paired with the functionality for parsing and modifying
security descriptors that was added in the security_descriptor_utils
file. For example, you can:
- change the owner of an object
- change the owning group of an object
- change ACEs for an object to add or remove permissions

The user and group versions support ADUser and ADGroup objects, as well
as lookup by name for user and group (sAMAccountName, dn, common name).
The object version supports ADObject objects and distinguished name strings.