Added new functions to add permissions to a user/group/object's security descriptor.
Instead of requiring that callers read a security descriptor, figure out modifications
themselves, and then set it back, we now have functions that allow a caller to pass
a user/group/object identifer, a list of SIDs to grant permissions to, and then
various lists of permission types to grant (access masks, rights guids, property
guids for read/write) and the library will handle reading the current security
descriptor, updating it as needed, and writing it back.

There are 3 functions, 1 for users, 1 for groups, and 1 for objects.
They accept inputs in the same forgiving formats that similar function
groups support.

Also added an enum for all AD rights guids. Property guids must still be specified
as strings, because there's a lot of properties and keeping up with all of them in
an enum seems very difficult, so tbd on whether that will be added.

Added a lot of type-hint enhancements for inputs with multiple possible types
because I learned about Union typing.