Added account takeover functionality for computers.
This commit fleshes out the function for taking over a computer account.
This allows users to specify a computer's sAMAccountName or distinguished
name, or the common name if it's unique, or specify an ADComputer or
ADObject object, and take over that computer.

The takeover consists of reading the object's encryption types and SPNs,
and then resetting/changing its password. The change of password ensures
others cannot impersonate this account, and the newly set password can
be used to generate kerberos keys.

Validation is done of the account control and encryption types to ensure
that they're supported.