Added functions to enable/disable accounts.
Added 2 new functions to enable or disable user/computer accounts,
which is an important part of account management. Often, automated
provisioning or deprovisioning systems may use enable/disable as
a precursor to deletion in order to allow for easier rollback if
it turns out that disabling the account caused a disruption.