Added user, group, and object attribute modification.
Added support for the following operations:
- Atomically append a value to an attribute
- Atomically append values for multiple attributes
- Atomically overwrite the value of an attribute
- Atomically overwrite the values of multiple attributes
for users, groups, and objects.

They support the standard specification of either one of the various
ADUser, ADGroup, etc. object or an appropriate string name that can
be found.

They support specifying attribute values as iterable types or as
singleton values, and the functions handle wrapping everything into
the proper LDAP format.

Also updated validation of passing in an ADObject object or an
appropriate string identifier for users, groups, and generic
objects.