Reworked computer management options.
Renamed the ad_computers.py file to managed_ad_objects.py and renamed
ad_users_and_groups.py to ad_objects.py

Descendants of ADObject represent objects in AD that we've read and
may make modifications to, but that we did not create and cannot
act as.
ManagedADObject and its descendents are objects that we can act as,
and have a password defined for them at a minimum (and kerberos
keys can be generated for them too).

This rename is technically a breaking change, but the library only
has like 5 users so that's ok and we should pull the trigger sooner
rather than later.

A bunch of computer management functions have also been added.
The same style of functions that existed for users and groups now
exist for computers. Find by various names/attributes, find groups
for computers, modify attributes in various ways, etc.
The type annotations for functions to change passwords and whatnot
have been expanded to include the new ADComputer object type used
in these functions.