Made session creation more flexible.
If someone specifies a SASL mechanism like KERBEROS as their authentication
mechanism, the library should automatically set the authentication
mechanism to SASL and pass that as the SASL mechanism.
A lot of people using AD don't really understand SASL vs. NTLM vs. SIMPLE
and just think of it as "kerberos vs. ntlm vs. simple", so we should
support that model of thought.