Added function for discovering trusted domains in a domain or using a session.
Added functions to look up the domains trusted by an AD Domain (or by the AD Domain
that a session is connected to) and return a list of ADTrustedDomain objects.
The session version of the function also supports a caching mechanism. Trust
information does not change often, so there's a cache of trusted domains in each
session object with a default lifetime of 24 hours.
The cache exists in the session as different users may have different rights on
different trust objects, so not all of them can see the same trusts.

Did some minor updates of various docstrings as well, and corrected how the library
raises errors due to permission/server issues so that we don't raise errors on
queries with no results.