Added helpers around netbios names and cross-domain users.
Added a function to find the netbios name from a domain, which can also be
called from a session. Also added a function for determining if a user
used for a session is from the domain that the session is communicating
with or not, which will be helpful when supporting automatic traversal of
cross-domain trusts vs. cross-forest trusts and such.

Fixed an issue with deepcopying trusted domains.
Updated __repr__ functions for domains and sessions.