Small enhancement to our function for finding a domain's netbios name
In a domain with subdomains, there may be partitions with different
netbios names. Previously, any of them could have been selected to use
for the domain's netbios name.

Now, we filter to only include partitions tied to our root dns name
for a domain, which should ignore subdomains and parent domains.