Added session transfer functionality amongst trusted domains.
Added a function to the ADTrustedDomain object to support transferring
sessions from the primary domain to be sessions for the trusted domain.
This transfers all connection attributes, including authentication
attributes, network attributes/strategies, and formatting-related
attributes, from the existing session and tries to authenticate a new
session with the trusted domain.
The function validates that the trusted domain trusts the primary domain
if the user is from the primary domain, and does some warning about
transitive trusts if the user appears to be from one, unless
validation is skipped.

A function also exists in a session to create transfer sessions to
all trusted domains. This allows a caller to spin out their sessions
and create new sessions. It could even be done recursively to get
sessions for all of the domains that are connected in the trust web
that a user is trusted within.