Added options to skip validation for peformance reasons.
Functions that do something special with users/groups/computers/objects
usually do validation to map names onto the appropriate AD Object,
to eventually reach distinguished names. This has the side effect of
validating existence.
However, when performing bulk operations on a lot of users/groups/etc.
this can result in a lot of queries.

This change adds a skip_validation option to nearly all of these functions,
allowing users who know that they're passing in valid distinguished names
to skip the lookup.
This can significantly improve performance in workflows that involve
bulk management of AD Objects where the same objects are reused a lot
over time and none are deleted.