Added windows path support.
Added support for finding objects by distinguished name (in an LDAP style)
or by windows canonical name (also called windows path style). This allows
users who don't know whether their location is a CN or an OU to look
it up and get the distinguished name of the location for use in other
operations, like joining a domain.

Some windows paths do not work because determining the correct escape
sequence for their special characters in filters is not yet working.