Added support for looking up the members of groups.
There are now functions to look up the members of a single group,
or the members of multiple groups, and attributes of the members.
These functions return members cast to the most accurate object
class based on information found about the member, and they may
be of type ADObject, ADUser, ADGroup, etc.