Added support for nested group member lookup.
Added functions to find the members of a group recursively and to
find the members of multiple groups recursively.
In Active Directory, groups can be members of other groups, and
so when finding the members of a group, you may wish to recurse
through those nested relationships.

When looking up recursive group memberships, one can specify the
depth to recurse, or leave it out (indicating to go until there
are no more nested groups).
The response will be a list of dictionaries, with the first mapping
the input group (or groups) to a list (or lists) of direct members.
The second dictionary takes the direct members and maps them to
their direct members. And so on and so forth.

When looking up recursive group memberships for a single group,
there's also an option to flatten the result and get a single
dictionary mapping the group to all of its members regardless of
depth, rather than its direct members.