Added support for configuring the local system for kerberos.
Added helpers for configuring krb5.conf (or the windows krb5.ini equivalent) or
a file of your choice for kerberos authentication for one or more AD domains.
This should make it easier to bootstrap a system into using kerberos programmatcally,
without needing to manually configure the file.

Options exist for setting the default domain for the local machine (e.g. a joined domain)
and for overwriting vs. merging with an existing configuration file.