Updated docs and added ADDomain with a new attribute for auto-configuration.
Added an option to ADDomain objects to auto-configure the local kerberos client
config file for the domain. This will update the relevant config file any
time the kerberos uris for the domain are changed.

Updated configuration logic to overwrite existing configurations for a realm
when updating the file and merging.

Updated docs.