Added managed user creation.
Added the ability to create users with a password and have kerberos
keys generated for them simultaneously.
The salt for AD users and AD computers is differently formed, with
users being closer to the traditional MIT Kerberos salt, while
computers are totally different. So updates were made there as well.