Added domain SID lookup and optimized primary group SID finding.
Added functionality for finding the SID of a domain via a session or an AD
domain object. The domain object will cache the SID since it doesn't change
in order to improve performance.

This also allows us to reduce the permissions needed to find the primary
group SIDs for a user, because we no longer need to read the objectSID of
the users in order to computer their primary group's SID, only the
primary group RID.