Added a UPN suffix lookup function.
This will return the valid user principal name suffixes for the domain.
These are domains that can be used for user principal names which impacts
user login.
UPN suffixes have use cases around zero-trust environments, domains
with a cloud/public internet component, and mergers and acquisitions
so being able to look them up is useful.

They're also important for multi-domain functionality in the future,
so this helps us to build towards that.